creates files for handlers and requests
