+++ lib/foods.js
@@ -1,18 +1,14 @@
 var createFood = require('./request')
 var handlers = require('./handlers')
+var requests = require('./request')
 
 $(document).ready(function() {
   loadFoods()
-  $('.foods-table').on('click', '.delete-row', deleteFood)
+  $('.foods-table').on('click', '.delete-row', requests.deleteFood())
 })
 
 function loadFoods() {
-  $.get('https://obscure-harbor-85447.herokuapp.com/api/v1/foods')
-  .then(function(foods) {
-    foods.forEach(function(food) {
-      handlers.appendFood(food)
-    })
-  })
+  requests.retrieveFoods()
 }
 
 
@@ -27,13 +23,13 @@ function updateFoodName(id, value) {
 function updateFoodCalories(id, value) {
 
 }
-function deleteFood(e) {
-  $.ajax({
-    type: 'DELETE',
-    url:'https://obscure-harbor-85447.herokuapp.com/api/v1/foods/' + e.target.id,
-    success: $(`.foods-table .row${e.target.id}`).remove()
-  })
-}
+// function deleteFood(e) {
+//   $.ajax({
+//     type: 'DELETE',
+//     url:'https://obscure-harbor-85447.herokuapp.com/api/v1/foods/' + e.target.id,
+//     success: $(`.foods-table .row${e.target.id}`).remove()
+//   })
+// }
 
 $('form').submit(function(e) {
   e.preventDefault()
