+++ lib/request.js
@@ -1,11 +1,11 @@
 var $ = require('jquery')
-var api = 'https://obscure-harbor-85447.herokuapp.com/api/v1'
+var api = 'https://obscure-harbor-85447.herokuapp.com/'
 var handlers = require('./handlers')
 
 function createFood(food, calories) {
   $.ajax({
     type: "POST",
-    url: api + "/foods",
+    url: api + "api/v1/foods",
     data: { food: {
       name: $.trim(food),
       calories: $.trim(calories)
@@ -19,4 +19,21 @@ function createFood(food, calories) {
   })
 }
 
-module.exports = createFood
+function retrieveFoods() {
+  $.get(api + '/api/v1/foods')
+  .then(function(foods) {
+    foods.forEach(function(food) {
+      handlers.appendFood(food)
+    })
+  })
+}
+
+function deleteFood(e) {
+  $.ajax({
+    type: 'DELETE',
+    url:'https://obscure-harbor-85447.herokuapp.com/api/v1/foods/' + e.target.id,
+    success: $(`.foods-table .row${e.target.id}`).remove()
+  })
+}
+
+module.exports = {createFood, retrieveFoods, deleteFood}
